+++ examples/flows/chat/promptflow-copilot/select_prompt.py
@@ -1,0 +1,9 @@
+from promptflow.core import tool
+
+
+@tool
+def my_python_tool(answer_question_prompt: str, refuse_prompt: str, not_relevant: bool) -> str:
+    if not_relevant:
+        return refuse_prompt
+
+    return answer_question_prompt