+++ examples/gen_test_data/gen_test_data/constants.py
@@ -1,0 +1,8 @@
+DOCUMENT_NODE = "document_node"
+TEXT_CHUNK = "text_chunk"
+NODES_FILE_NAME = "document_nodes.jsonl"
+DETAILS_FILE_NAME = "test-data-gen-details.jsonl"
+PARALLEL_RUN_STEP_FILE_NAME = "parallel_run_step.jsonl"
+SUMMARY_FILE_NAME = "test-data-gen-summary.json"
+TEST_DATA_FILE_NAME = "test-data.jsonl"
+SUPPORT_FILE_TYPE = [".docx", ".pdf", ".ipynb", ".md", ".txt"]