+++ src/promptflow-devkit/tests/sdk_cli_test/unittests/test_utils.py
@@ -33,7 +33,7 @@ from promptflow._constants import LAST_CHECK_TIME, PF_VERSION_CHECK
 from promptflow._sdk._constants import HOME_PROMPT_FLOW_DIR, PROMPT_FLOW_HOME_DIR_ENV_VAR
 from promptflow._sdk._errors import GenerateFlowToolsJsonError
 from promptflow._sdk._telemetry.logging_handler import get_scrubbed_cloud_role
-from promptflow._sdk._utils import (
+from promptflow._sdk._utilities.general_utils import (
     _generate_connections_dir,
     decrypt_secret_value,
     encrypt_secret_value,
@@ -42,6 +42,7 @@ from promptflow._sdk._utils import (
     get_mac_address,
     get_system_info,
     refresh_connections_dir,
+    resolve_flow_language,
 )
 from promptflow._sdk._version_hint_utils import check_latest_version
 from promptflow._utils.load_data import load_data
@@ -51,6 +52,7 @@ from promptflow.core._utils import (
     override_connection_config_with_environment_variable,
     resolve_connections_environment_variable_reference,
 )
+from promptflow.exceptions import UserErrorException
 
 TEST_ROOT = PROMPTFLOW_ROOT / "tests"
 CONNECTION_ROOT = TEST_ROOT / "test_configs/connections"
@@ -82,7 +84,7 @@ class TestUtils:
             mock_sqlite_op()
         # assert function execution time from stdout
         out, _ = capfd.readouterr()
-        assert out.count("sqlite op...") == 3
+        assert out.count("sqlite op...") <= 10
 
     def test_resolve_connections_environment_variable_reference(self):
         connections = {
@@ -335,6 +337,43 @@ class TestUtils:
             importlib.reload(_constants)
             assert _constants.HOME_PROMPT_FLOW_DIR.as_posix() == (Path.home() / ".promptflow").resolve().as_posix()
         importlib.reload(_constants)
+
+    def test_resolve_flow_language(self):
+        # dag flow
+        lan = resolve_flow_language(flow_path=TEST_ROOT / "test_configs" / "flows" / "csharp_flow")
+        assert lan == "csharp"
+
+        lan = resolve_flow_language(flow_path=TEST_ROOT / "test_configs" / "flows" / "chat_flow")
+        assert lan == "python"
+
+        # flex flow
+        lan = resolve_flow_language(flow_path=TEST_ROOT / "test_configs" / "eager_flows" / "basic_callable_class")
+        assert lan == "python"
+
+        lan = resolve_flow_language(
+            flow_path=TEST_ROOT / "test_configs" / "eager_flows" / "basic_dummy_csharp_flex_flow"
+        )
+        assert lan == "csharp"
+
+        # prompty
+        lan = resolve_flow_language(flow_path=TEST_ROOT / "test_configs" / "prompty" / "prompty_example.prompty")
+        assert lan == "python"
+
+        with pytest.raises(UserErrorException) as ex:
+            resolve_flow_language()
+        assert "Either flow_path or yaml_dict should be provided." in ex.value.message
+
+        with pytest.raises(UserErrorException) as ex:
+            resolve_flow_language()
+        assert "Either flow_path or yaml_dict should be provided." in ex.value.message
+
+        with pytest.raises(UserErrorException) as ex:
+            resolve_flow_language(flow_path="mock_path", yaml_dict="mock_dict")
+        assert "Only one of flow_path and yaml_dict should be provided." in ex.value.message
+
+        with pytest.raises(UserErrorException) as ex:
+            resolve_flow_language(flow_path="mock_path")
+        assert "must exist and of suffix yaml, yml or prompty." in ex.value.message
 
 
 @pytest.mark.unittest